This change contains a bunch of enhancements to the tests and test framework. The biggest part of it is allowing multithreaded upload of files and validation of transfers. I wont go into details about the test changes though. (#157)
As for the real changes:
I moved the max connections/sessions out to external configuration for both the api and the worker (they are independent settings). In retrospect maybe I could have called them the same thing - we can talk about it during the review.

Changed logback settings for the sshpool logs to go with the changes in the shared library.

Made changes in the SSHDataClient to prevent deadlocks.

+++ api/src/test/java/edu/utexas/tacc/tapis/files/integration/transfers/configs/UploadFilesConfig.java
@@ -1,14 +1,12 @@
 package edu.utexas.tacc.tapis.files.integration.transfers.configs;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-
 import java.nio.file.Path;
 
 public class UploadFilesConfig {
     private String uploadSystem;
     private Path uploadPath;
     private int count;
-    private long size;
+    private int size;
 
     public String getUploadSystem() {
         return uploadSystem;
@@ -22,7 +20,7 @@ public class UploadFilesConfig {
         return count;
     }
 
-    public long getSize() {
+    public int getSize() {
         return size;
     }
 }
